+++ sdk/internal/recording/recording_test.go
@@ -518,10 +518,10 @@ func TestLiveOnly(t *testing.T) {
 
 func TestHostAndScheme(t *testing.T) {
 	r := RecordingOptions{UseHTTPS: true}
-	require.Equal(t, r.Host(), "https")
-	require.Equal(t, r.Scheme(), "localhost:5001")
+	require.Equal(t, r.scheme(), "https")
+	require.Equal(t, r.hostAndPort(), "localhost:5001")
 
 	r.UseHTTPS = false
-	require.Equal(t, r.Host(), "http")
-	require.Equal(t, r.Scheme(), "localhost:5000")
+	require.Equal(t, r.scheme(), "http")
+	require.Equal(t, r.hostAndPort(), "localhost:5000")
 }